fix(cri): match Darwin ioctl request type explicitly
Actions caught E0308 before runtime tests: the pinned libc exposes TIOCSCTTY as u32 and ioctl requests as c_ulong. Preserve every test and native check; cast the request, not its result.

+++ crates/lightr-cri-backend/src/stream_tty.rs
@@ -17,7 +17,7 @@ pub(super) unsafe fn configure_child() -> io::Result<()> {
     if unsafe { libc::setsid() } < 0 {
         return Err(io::Error::last_os_error());
     }
-    if unsafe { libc::ioctl(libc::STDIN_FILENO, libc::TIOCSCTTY, 0) } < 0 {
+    if unsafe { libc::ioctl(libc::STDIN_FILENO, libc::TIOCSCTTY as libc::c_ulong, 0) } < 0 {
         return Err(io::Error::last_os_error());
     }
     let terminal = unsafe {
